requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
statsmodels>=0.13.0
matplotlib>=3.4.0
seaborn>=0.11.0

# LLM dependencies
openai>=1.0.0
python-dotenv>=0.19.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Development
black>=22.0.0
flake8>=4.0.0
isort>=5.0.0
mypy>=0.910
pre-commit>=2.0.0

# Type hints
typing-extensions>=4.0.0 